botansik_dugnad.yaml
Dear Christian, 

I have done my best trying to update the code with the specifications for the Flora Norvegica dugnad's project. I have almost no experience in programming, so I have done so intuitively and I still have many doubts. I really hope I have not made a mess.

I will now send an email with the image file and the pdf. It would be nice to add an "Instruksjon" file and an "Instruksjonvideo" in the future. For now I have added the instructions in the "help: / introduction:" functions, but I am not sure it is the right place (and I don't know how to put the intros in the text). 

Please, do let me know if I can improve this in any way or if something is not clear. 
Don't hesitate to contact me at irene.teixidor.toneu@gmail.com or i.t.toneu@nhm.uio.no
My best wishes,
Irene